API folder added
Added the API files to a folder for moving to the main branch
