+++ API/Models/LobbyContext.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace QuizAPI.Models
+{
+	public class LobbyContext : DbContext
+	{
+		public LobbyContext()
+		{
+		}
+
+		public LobbyContext(DbContextOptions<LobbyContext> options)
+			: base(options)
+		{
+		}
+
+		public DbSet<Lobby> LobbyItems { get; set; }
+	}
+}
