+++ API/README.md
@@ -1,0 +1,3 @@
+# API
+
+Contains the API code, which uses ASP.NET in C#, as well as documentation and files related to the API.
