--- Models/LobbyContext.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace QuizAPI.Models
-{
-	public class LobbyContext : DbContext
-	{
-		public LobbyContext()
-		{
-		}
-
-		public LobbyContext(DbContextOptions<LobbyContext> options)
-			: base(options)
-		{
-		}
-
-		public DbSet<Lobby> LobbyItems { get; set; }
-	}
-}
